#98: Consolidate libs + test fixtures
