+++ audio/al/build.gradle.kts
@@ -8,12 +8,8 @@ apply(from = "$rootDir/publish.gradle.kts")
 description = "OpenAL audio system and struct implementations"
 
 dependencies {
-	implementation(libs.bundles.stdlib)
-
 	api(projects.platform)
 	implementation(projects.core)
-
-	testImplementation(libs.bundles.test)
 }
 
 (extra["setupLwjgl"] as (Any) -> Unit)(listOf(libs.lwjgl.openal))
