+++ core/src/test/groovy/dev/kkorolyov/pancake/core/action/TransformActionSpec.groovy
@@ -9,7 +9,7 @@ import dev.kkorolyov.pancake.platform.math.Vector3
 
 import spock.lang.Specification
 
-import static dev.kkorolyov.pancake.platform.SpecUtilities.randVector
+import static dev.kkorolyov.pancake.test.SpecUtilities.randVector
 
 class TransformActionSpec extends Specification {
 	Vector3 setPosition = randVector()
