--- editor/core/src/test/java/test/Test.java
@@ -1,7 +1,0 @@
-package test;
-
-public class Test {
-	public static void main(String[] args) {
-		System.out.println("beep");
-	}
-}
