+++ graphics/build.gradle.kts
@@ -8,10 +8,6 @@ apply(from = "$rootDir/publish.gradle.kts")
 description = "Common rendering system utilities"
 
 dependencies {
-	implementation(libs.bundles.stdlib)
-
 	api(projects.platform)
 	implementation(projects.core)
-
-	testImplementation(libs.bundles.test)
 }
