+++ input/glfw/build.gradle.kts
@@ -8,13 +8,9 @@ apply(from = "$rootDir/publish.gradle.kts")
 description = "GLFW input system and control implementations"
 
 dependencies {
-	implementation(libs.bundles.stdlib)
-
 	api(projects.platform)
 	api(projects.input)
 	implementation(projects.core)
-
-	testImplementation(libs.bundles.test)
 }
 
 (extra["setupLwjgl"] as (Any) -> Unit)(listOf(libs.lwjgl.glfw))
