+++ platform/src/testFixtures/groovy/dev/kkorolyov/pancake/test/SpecUtilities.groovy
@@ -1,4 +1,4 @@
-package dev.kkorolyov.pancake.platform
+package dev.kkorolyov.pancake.test
 
 import dev.kkorolyov.pancake.platform.math.Vector3
