--- test-utils/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-	groovy
-}
-
-description = "Test utilities"
-
-repositories {
-	mavenCentral()
-}
-dependencies {
-	implementation(libs.spock)
-	implementation(projects.platform)
-}
